requirements.txt
fastapi
uvicorn[standard]
motor
pymongo
passlib[bcrypt]
bcrypt==4.0.1
email-validator
python-jose[cryptography]
httpx
pydantic-settings
requests